ref: externalized `CourseContentDialog` into component
This way the component rendering the course-content dialog for creating / editing a given course-content is easier to maintain and re-usable.
